Add base wallet account
