+++ cuenca/resources/savings.py
@@ -1,28 +1,34 @@
 from dataclasses import dataclass
+from datetime import datetime
 from typing import ClassVar, cast
 
-from cuenca_validations.types import Currency, SavingRequest
+from cuenca_validations.types import Currency, SavingCategory, SavingRequest
 
-from .base import Creatable, Updateable
-from .resources import Wallet
+from .base import Creatable, Updateable, Wallet
 
 
 @dataclass
 class Saving(Wallet, Creatable, Updateable):
     _resource: ClassVar = 'Saving'
 
-    category: str
+    amount: int
+    category: SavingCategory
+    end_date: datetime
     name: str
-    goal: int
-    end_goal: int
 
     @classmethod
-    def create(cls, name: str, category: str, goal: int, end_goal: int):
+    def create(
+        cls,
+        name: str,
+        amount: int,
+        end_date: datetime,
+        category: SavingCategory,
+    ):
         request = SavingRequest(
+            amount=amount,
             name=name,
             category=category,
-            goal=goal,
-            end_goal=end_goal,
+            end_date=end_date,
             currency=Currency.mxn,  # For now
         )
         return cast('Saving', cls._create(**request.dict()))
